MemoryDomainProxy
