+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -100,6 +100,7 @@
     <Compile Include="EventWarlock\WarlockAction.cs" />
     <Compile Include="EventWarlock\WarlockActions\WarlockActionEcho.cs" />
     <Compile Include="EventWarlock\WarlockConditions\FirstEqualsSecond.cs" />
+    <Compile Include="MemoryDomains\MemoryDomainProxy.cs" />
     <Compile Include="MemoryDomains\MemoryInterface.cs" />
     <Compile Include="MemoryDomains\VirtualMemoryDomain.cs" />
     <Compile Include="IRealTime.cs" />
